% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOfisher-package.R
\docType{package}
\name{GOfisher-package}
\alias{GOfisher}
\alias{GOfisher-package}
\title{GOfisher: Simple GO enrichment analysis using the Fisher test}
\description{
The GOfisher package includes a sort of functions to perform Fisher's exact probability tests with FDR correction.
}
\details{
The original source codes are available at Nagano Lab's GitHub repository (\url{https://github.com/naganolab/Lasy-Seq/tree/master/TempResponse_A_thaliana}).
}
\references{
\itemize{
\item Kamitani M, Kashima M, Tezuka A, Nagano AJ. (2019) Lasy-Seq: A High-Throughput Library Preparation Method for RNA-Seq and Its Application in the Analysis of Plant Responses to Fluctuating Temperatures. Scientific Reports 9:7091. \url{https://doi.org/10.1038/s41598-019-43600-0.}
}
}
\author{
A.J. Nagano & Y. Sato
}
\keyword{internal}
