% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng.GetGOTerms.R
\name{ng.GetGOTerms}
\alias{ng.GetGOTerms}
\title{get GO terms}
\usage{
ng.GetGOTerms(GOid)
}
\arguments{
\item{GOid}{input GO ID}
}
\description{
Function to get GO term from GO ID
}
\author{
A.J. Nagano & Y. Sato
}
