% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng.GOfisher2REVIGO.R
\name{ng.GOfisher2REVIGO}
\alias{ng.GOfisher2REVIGO}
\title{Converting GOfisher outputs into REVIGO inputs}
\usage{
ng.GOfisher2REVIGO(r, gn.test, cgt)
}
\arguments{
\item{r}{output table from \code{ng.mft()}}

\item{gn.test}{a set of genes to be examined}

\item{cgt}{output from \code{ng.GOprep_TAIR()}, \code{[,1]:"locus", [,2]:"GOid"}}
}
\value{
A Data Frame object useful for the "rrvigo" package
}
\description{
This is an utility function to convert the \code{ng.mft()} output into the input of the "rrvigo" package.
}
\examples{
data(ulg)
data(gl)
fisher.res <- ng.mft(ulg, gl)
input <- ng.GOfisher2REVIGO(fisher.res[fisher.res[,"xtt"]>1,], gl, ulg)
}
\seealso{
\code{\link[=ng.mft]{ng.mft()}}
}
\author{
Y. Sato
}
