% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng.GOprep_TAIR.R
\name{ng.GOprep_TAIR}
\alias{ng.GOprep_TAIR}
\title{Preparing a list of GO terms with TAIR annotation files}
\usage{
ng.GOprep_TAIR(fn)
}
\arguments{
\item{fn}{path to the file name. The input file follows TAIR file format, named "ATH_GO_GOSLIM.txt".}
}
\value{
A matrix listing AGI codes and GO terms
}
\description{
Making GO list from TAIR input file
}
\seealso{
\code{\link[=ng.mft]{ng.mft()}}
}
\author{
A.J. Nagano & Y. Sato
}
