% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng.mft.R
\name{ng.mft}
\alias{ng.mft}
\title{Multiple Fisher tests}
\usage{
ng.mft(cgt, gn.test, alternative = "greater")
}
\arguments{
\item{cgt}{output from \code{ng.GOprep_TAIR()}, \code{[,1]:"locus", [,2]:"GOid"}}

\item{gn.test}{contig names for testing}

\item{alternative}{alternative hypothesis for the Fisher test}
}
\value{
No. of genes, p.value, and calculation time
}
\description{
Perform Fisher's exact probability tests multiple times
}
\examples{
data(ulg)
head(ulg)
data(gl)
head(gl)
fisher.res <- ng.mft(ulg, gl)
head(fisher.res)
# The numbers of genes in the output table are given as follows.
#               gn.test
#             TRUE FALSE
# cgt   TRUE   xtt   xft   xnt
#      FALSE   xtf   xff   xnf
#              xtn   xfn   xnn
}
\seealso{
\code{\link[=ng.GOprep_TAIR]{ng.GOprep_TAIR()}}
}
\author{
A.J. Nagano & Y. Sato
}
