% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng.prepGOtestOutTable.R
\name{ng.prepGOtestOutTable}
\alias{ng.prepGOtestOutTable}
\title{Preparing an output table for GO tests}
\usage{
ng.prepGOtestOutTable(r, adj.method = "BH", alpha = 0.01)
}
\arguments{
\item{r}{output table from \code{ng.mft()}}

\item{adj.method}{A method for the correction of multiple testing. Passed to \code{p.adjust()}}

\item{alpha}{cut-off value of the false discovery rate}
}
\description{
Result table for Fisher tests, with multiple testing corrected
}
\examples{
data(ulg)
data(gl)
fisher.res <- ng.mft(ulg, gl)
GO.list <- ng.prepGOtestOutTable(fisher.res[fisher.res[,"xtt"]>1,], alpha=0.05)
GO.list[order(GO.list[,1]),]
}
\seealso{
\code{\link[=ng.mft]{ng.mft()}} \code{\link[=p.adjust]{p.adjust()}}
}
\author{
A.J. Nagano & Y. Sato
}
