% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng.SearchGOTerms.R
\name{ng.SearchGOTerms}
\alias{ng.SearchGOTerms}
\title{search GO terms}
\usage{
ng.SearchGOTerms(gn.test, GOid, cgt)
}
\arguments{
\item{gn.test}{a set of genes to be examined}

\item{GOid}{query GO ID}

\item{cgt}{output from \code{ng.GOprep_TAIR()}, \code{[,1]:"locus", [,2]:"GOid"}}
}
\description{
Function to search genes having a given GO ID
}
\examples{
data(ulg)
data(gl)
res <- ng.SearchGOTerms(gl, "GO:0006952", ulg)

data(des)
des[res,]
}
\author{
A.J. Nagano & Y. Sato
}
